"[Feat/#week2] Refactor"
